Update django_svelte.py
Since things still need to be served out of the `STATIC_ROOT` / under the `STATIC_URL`, I believe users (like me)
will want to supply the Svelte path to the resource, as opposed to the one Django uses in its static file server.
This adjustment accounts for easing the migration from Svelte-land to Django-land.